Added DataGroupingSeriesComposition.
