+++ ts/Extensions/DataGrouping/ApproximationRegistry.ts
@@ -24,6 +24,19 @@ import type { ApproximationTypeRegistry } from './ApproximationType';
  *
  * */
 
+/**
+ * Define the available approximation types. The data grouping
+ * approximations takes an array or numbers as the first parameter. In case
+ * of ohlc, four arrays are sent in as four parameters. Each array consists
+ * only of numbers. In case null values belong to the group, the property
+ * .hasNulls will be set to true on the array.
+ *
+ * @product highstock
+ *
+ * @private
+ * @name Highcharts.approximations
+ * @type {Highcharts.Dictionary<Function>}
+ */
 const ApproximationRegistry = {
     // approximations added programmatically
 } as ApproximationTypeRegistry;
