+++ ts/masters/modules/datagrouping.src.ts
@@ -12,7 +12,11 @@
 'use strict';
 import Highcharts from '../../Core/Globals.js';
 import '../../Extensions/DataGrouping/ApproximationDefaults.js';
-import dataGrouping from '../../Extensions/DataGrouping.js';
+import ApproximationRegistry from '../../Extensions/DataGrouping/ApproximationRegistry.js';
+import DataGrouping from '../../Extensions/DataGrouping/DataGrouping.js';
 const G: AnyRecord = Highcharts;
-G.dataGrouping = dataGrouping;
-export default dataGrouping;
+G.dataGrouping = {
+    approximations: ApproximationRegistry
+};
+DataGrouping.compose(G.Axis, G.Series, G.Tooltip);
+export default G.dataGrouping;
